Remove two screws and make it screw togheter
By removing two screws I was able to make the whole keyboard stick
togheter with the help of screws in the middle. By placing them in the
exact middle section we have overlap between the parts that can be
screwed togheter.

Also made sure that the planned for sccrews actually reach all the way
in. You will need 3 M3x20 and 3 M3x10 to asseble this.